pc/requirements.txt
pyserial
pyautogui

install with 

pip install -r pc/requirements.txt